Update and rename README.py to drone_route.py
This was my first assessment to validate my skills as a python programmer. It took me approximately three weeks to complete in my spare time whilst working a full time job. The assessment was given to me as an incentive to get those in my line of work more 'tech savvy' and I have now developed a further understanding and interest in python programming. 

Please note this will not run without the correct '.txt' files located in the same folder. This is here solely for feedback on my use of syntax and methods.

Appreciate it, much love. 

Chopper